Move `cmd/common` to `shared`
